Add `django-uuidfield` requirement in `setup.py`.
For django versions < 1.8, this package depends on `django-uuidfield`
as it is used in the models, however it is not listed in the setup.py
file, so when one does a `pip install` of this package in their Django
project they end up with an incomplete and broken installation, and has
to hunt down where to find the uuidfield module.  This change adds the
`django-uuidfield` package in the `install_requires` of this package's
installer.